# openapi.yaml
openapi: 3.1.0
info:
  title: SchemaSprint BFF API
  version: 0.1.0
  description: Versioned same-origin contract for the private Web/PWA MVP. Provider traces and secrets are never returned.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /v1/policies:
    get:
      tags: [Account]
      operationId: listRequiredPolicies
      responses:
        '200':
          description: Latest effective policies per kind, readable before acknowledgement.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Policy'}}}}
        default: {$ref: '#/components/responses/Error'}
  /v1/profile:
    get:
      tags: [Account]
      operationId: getProfile
      responses:
        '200': {description: Profile., content: {application/json: {schema: {$ref: '#/components/schemas/Profile'}}}}
        default: {$ref: '#/components/responses/Error'}
    patch:
      tags: [Account]
      operationId: updateProfile
      parameters: [{$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                displayName: {type: string, minLength: 1, maxLength: 64}
                locale: {$ref: '#/components/schemas/Locale'}
                theme: {type: string, enum: [system, light, dark]}
      responses:
        '200': {description: Updated profile., content: {application/json: {schema: {$ref: '#/components/schemas/Profile'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/problems:
    get:
      tags: [Problems]
      operationId: listProblems
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
        - {name: q, in: query, schema: {type: string, maxLength: 200}}
        - {name: difficulty, in: query, schema: {$ref: '#/components/schemas/Difficulty'}}
        - {name: genre, in: query, schema: {type: string, maxLength: 64}}
        - {name: inputFormat, in: query, schema: {$ref: '#/components/schemas/InputFormat'}}
        - {name: status, in: query, schema: {type: string, enum: [passed, not_passed, unsubmitted]}}
        - {name: publishedFrom, in: query, schema: {type: string, format: date}}
        - {name: publishedTo, in: query, schema: {type: string, format: date}}
        - {$ref: '#/components/parameters/Locale'}
      responses:
        '200': {description: Published library., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemPage'}}}}
        default: {$ref: '#/components/responses/Error'}
  /v1/problems/{problemId}:
    get:
      tags: [Problems]
      operationId: getProblem
      parameters:
        - {$ref: '#/components/parameters/ProblemId'}
        - {$ref: '#/components/parameters/Locale'}
        - {name: version, in: query, schema: {type: integer, minimum: 1}}
      responses:
        '200': {description: "Published statement, official answer and explanation; opening is penalty-free.", content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}}
        default: {$ref: '#/components/responses/Error'}
  /v1/problems/{problemId}/draft:
    get:
      tags: [Workspace]
      operationId: getServerDraft
      parameters: [{$ref: '#/components/parameters/ProblemId'}]
      responses:
        '200': {description: Premium synchronized draft., content: {application/json: {schema: {$ref: '#/components/schemas/Draft'}}}}
        default: {$ref: '#/components/responses/Error'}
    put:
      tags: [Workspace]
      operationId: putServerDraft
      parameters:
        - {$ref: '#/components/parameters/ProblemId'}
        - {$ref: '#/components/parameters/CsrfToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {name: If-Match, in: header, required: true, schema: {type: string, pattern: '^"[1-9][0-9]*"$'}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DraftWrite'}}}}
      responses:
        '200': {description: Saved with incremented revision., content: {application/json: {schema: {$ref: '#/components/schemas/Draft'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Workspace]
      operationId: deleteServerDraft
      parameters:
        - {$ref: '#/components/parameters/ProblemId'}
        - {$ref: '#/components/parameters/CsrfToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {name: If-Match, in: header, required: true, schema: {type: string, pattern: '^"[1-9][0-9]*"$'}}
      responses:
        '204': {description: Server draft deleted.}
        default: {$ref: '#/components/responses/Error'}
  /v1/submissions:
    post:
      tags: [Assessment]
      operationId: createSubmission
      parameters: [{$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubmissionCreate'}}}}
      responses:
        '202': {description: Immutable snapshot accepted; semantic assessment queued., content: {application/json: {schema: {$ref: '#/components/schemas/Submission'}}}}
        '200': {description: Idempotent replay returns the original resource., content: {application/json: {schema: {$ref: '#/components/schemas/Submission'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
        default: {$ref: '#/components/responses/Error'}
  /v1/submissions/{submissionId}:
    get:
      tags: [Assessment]
      operationId: getSubmission
      parameters: [{$ref: '#/components/parameters/SubmissionId'}]
      responses:
        '200': {description: Owned immutable submission and current assessment state., content: {application/json: {schema: {$ref: '#/components/schemas/Submission'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/dashboard:
    get:
      tags: [Assessment]
      operationId: getDashboard
      parameters:
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      responses:
        '200': {description: Unique pass totals/calendar only; no accuracy metric., content: {application/json: {schema: {$ref: '#/components/schemas/Dashboard'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/notices:
    get:
      tags: [Notices]
      operationId: listNotices
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Locale'}]
      responses:
        '200': {description: Approved archive and unread state., content: {application/json: {schema: {$ref: '#/components/schemas/NoticePage'}}}}
        default: {$ref: '#/components/responses/Error'}
  /v1/notices/{noticeVersionId}/read:
    put:
      tags: [Notices]
      operationId: markNoticeRead
      parameters:
        - {name: noticeVersionId, in: path, required: true, schema: {$ref: '#/components/schemas/Ulid'}}
        - {$ref: '#/components/parameters/CsrfToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '204': {description: Read marker recorded idempotently.}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: __Host-schemasprint_session}
  parameters:
...
    CsrfToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
...
    ProblemId: {name: problemId, in: path, required: true, schema: {$ref: '#/components/schemas/Ulid'}}
    SubmissionId: {name: submissionId, in: path, required: true, schema: {$ref: '#/components/schemas/Ulid'}}
...
    Locale: {name: locale, in: query, schema: {$ref: '#/components/schemas/Locale'}}
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 30}}
  responses:
    Error:
      description: Structured safe error.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemError'}}}
    Conflict:
      description: Idempotency payload or optimistic-revision conflict.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemError'}}}
  schemas:
    Ulid: {type: string, pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'}
    Locale: {type: string, enum: [en, ja, zh-CN, ko, es, pt-BR]}
    Difficulty: {type: string, enum: [easy, mid, hard]}
    InputFormat: {type: string, enum: [gui, mermaid, ddl, dbml, mixed]}
...
    Profile:
      type: object
      additionalProperties: false
      required: [id, displayName, locale, theme, roles, entitlement]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        displayName: {type: string}
        locale: {$ref: '#/components/schemas/Locale'}
        theme: {type: string, enum: [system, light, dark]}
        roles: {type: array, uniqueItems: true, items: {type: string, enum: [learner, moderator, admin, owner]}}
        entitlement: {$ref: '#/components/schemas/Entitlement'}
    Entitlement:
      type: object
      additionalProperties: false
      required: [premium, billingEnabled, adsEnabled]
      properties:
        premium: {type: boolean}
        validUntil: {type: [string, 'null'], format: date-time}
        billingEnabled: {type: boolean, const: false}
        adsEnabled: {type: boolean, const: false}
    Policy:
      type: object
      additionalProperties: false
      required: [id, kind, version, content, requiredFrom, acknowledged]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        kind: {type: string, enum: [terms, privacy, community]}
        version: {type: integer, minimum: 1}
        content: {type: string}
        requiredFrom: {type: string, format: date-time}
        acknowledged: {type: boolean}
    ProblemSummary:
      type: object
      additionalProperties: false
      required: [id, version, title, difficulty, genre, inputFormat, tags, publicationDate, passed, submitted]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        version: {type: integer, minimum: 1}
        title: {type: string}
        difficulty: {$ref: '#/components/schemas/Difficulty'}
        genre: {type: string}
        inputFormat: {$ref: '#/components/schemas/InputFormat'}
        tags: {type: array, items: {type: string}}
        publicationDate: {type: string, format: date}
        passed: {type: boolean}
        submitted: {type: boolean}
        translationStatus: {type: string, enum: [source, translated, pending, unavailable]}
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [id, version, title, difficulty, genre, inputFormat, tags, publicationDate, passed, submitted, translationStatus, problemVersionId, statement, officialAnswer, explanation, sourceLanguage, requestedLocale, rubricSummary]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        version: {type: integer, minimum: 1}
        title: {type: string}
        difficulty: {$ref: '#/components/schemas/Difficulty'}
        genre: {type: string}
        inputFormat: {$ref: '#/components/schemas/InputFormat'}
        tags: {type: array, items: {type: string}}
        publicationDate: {type: string, format: date}
        passed: {type: boolean}
        submitted: {type: boolean}
        translationStatus: {type: string, enum: [source, translated, pending, unavailable]}
        problemVersionId: {$ref: '#/components/schemas/Ulid'}
        statement: {type: string}
        officialAnswer: {$ref: '#/components/schemas/CanonicalSchema'}
        explanation: {type: string}
        sourceLanguage: {type: string, const: en}
        requestedLocale: {$ref: '#/components/schemas/Locale'}
        rubricSummary:
          type: array
          maxItems: 2000
          items:
            type: object
            additionalProperties: false
            required: [id, description, weight, critical, implicit]
            properties:
              id: {$ref: '#/components/schemas/Ulid'}
              description: {type: string}
              weight: {type: integer, minimum: 1, maximum: 1000000}
              critical: {type: boolean}
              implicit: {type: boolean}
    ProblemPage:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ProblemSummary'}}
        nextCursor: {type: [string, 'null']}
    CanonicalSchema:
      type: object
      additionalProperties: false
      required: [schemaVersion, tables, relationships]
      properties:
        schemaVersion: {type: integer, const: 1}
        tables:
          type: array
          maxItems: 200
          items: {$ref: '#/components/schemas/SchemaTable'}
        relationships:
          type: array
          maxItems: 2000
          items: {$ref: '#/components/schemas/SchemaRelationship'}
        enums: {type: array, maxItems: 200, items: {type: object}}
        views: {type: array, maxItems: 200, items: {type: object}}
        triggers: {type: array, maxItems: 200, items: {type: object}}
        policies: {type: array, maxItems: 500, items: {type: object}}
        partitions: {type: array, maxItems: 200, items: {type: object}}
    SchemaTable:
      type: object
      additionalProperties: false
      required: [id, name, columns, indexes, checks, position]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 128}
        columns:
          type: array
          maxItems: 200
          items:
            type: object
            additionalProperties: false
            required: [id, name, dataType, nullable]
            properties:
              id: {type: string, minLength: 1, maxLength: 128}
              name: {type: string, minLength: 1, maxLength: 128}
              dataType: {type: string, minLength: 1, maxLength: 256}
              nullable: {type: boolean}
              primaryKey: {type: boolean, default: false}
              defaultExpression: {type: [string, 'null'], maxLength: 2000}
              generatedExpression: {type: [string, 'null'], maxLength: 2000}
        indexes: {type: array, maxItems: 200, items: {type: object}}
        checks: {type: array, maxItems: 200, items: {type: object}}
        position:
          type: object
          required: [x, y]
          properties: {x: {type: number}, y: {type: number}}
    SchemaRelationship:
      type: object
      additionalProperties: false
      required: [id, fromTableId, fromColumnIds, toTableId, toColumnIds, fromCardinality, toCardinality]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        fromTableId: {type: string}
        fromColumnIds: {type: array, minItems: 1, maxItems: 32, items: {type: string}}
        toTableId: {type: string}
        toColumnIds: {type: array, minItems: 1, maxItems: 32, items: {type: string}}
        fromCardinality: {type: string, enum: ['0..1', '1', '0..*', '1..*']}
        toCardinality: {type: string, enum: ['0..1', '1', '0..*', '1..*']}
        onDelete: {type: string, enum: [no_action, restrict, cascade, set_null, set_default]}
    EditorSources:
      type: object
      additionalProperties: false
      properties:
        ddl: {type: string, maxLength: 1048576}
        mermaid: {type: string, maxLength: 1048576}
        dbml: {type: string, maxLength: 1048576}
        activeFormat: {$ref: '#/components/schemas/InputFormat'}
        warnings: {type: array, maxItems: 500, items: {type: string}}
    DraftWrite:
      type: object
      additionalProperties: false
      required: [canonicalSchema, editorSources, layout]
      properties:
        canonicalSchema: {$ref: '#/components/schemas/CanonicalSchema'}
        editorSources: {$ref: '#/components/schemas/EditorSources'}
        layout: {type: object}
    Draft:
      allOf:
        - {$ref: '#/components/schemas/DraftWrite'}
        - type: object
          required: [problemId, revision, updatedAt]
          properties:
            problemId: {$ref: '#/components/schemas/Ulid'}
            revision: {type: integer, minimum: 1}
            updatedAt: {type: string, format: date-time}
    SubmissionCreate:
      type: object
      additionalProperties: false
      required: [problemId, problemVersionId, canonicalSchema, editorSources, staticAnalysis]
      properties:
        problemId: {$ref: '#/components/schemas/Ulid'}
        problemVersionId: {$ref: '#/components/schemas/Ulid'}
        canonicalSchema: {$ref: '#/components/schemas/CanonicalSchema'}
        editorSources: {$ref: '#/components/schemas/EditorSources'}
        staticAnalysis:
          type: object
          required: [parserVersion, diagnostics]
          properties:
            parserVersion: {type: string, minLength: 1, maxLength: 100}
            diagnostics: {type: array, maxItems: 2000, items: {type: object}}
    Submission:
      type: object
      additionalProperties: false
      required: [id, problemId, problemVersionId, state, canonicalSchema, createdAt]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        problemId: {$ref: '#/components/schemas/Ulid'}
        problemVersionId: {$ref: '#/components/schemas/Ulid'}
        state: {type: string, enum: [pending, assessed, failed]}
        canonicalSchema: {$ref: '#/components/schemas/CanonicalSchema'}
        result: {anyOf: [{$ref: '#/components/schemas/AssessmentResult'}, {type: 'null'}]}
        failureCode: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
    AssessmentResult:
      type: object
      additionalProperties: false
      required: [score, exactFull, passed, contradiction, confidence, requirements]
      properties:
        score: {type: number, minimum: 0, maximum: 100, multipleOf: 0.1}
        exactFull: {type: boolean}
        passed: {type: boolean}
        contradiction: {type: boolean}
        confidence: {type: number, minimum: 0, maximum: 100}
        requirements:
          type: array
          maxItems: 2000
          items:
            type: object
            required: [rubricItemId, decision, source, impact]
            properties:
              rubricItemId: {$ref: '#/components/schemas/Ulid'}
              decision: {type: string, enum: [satisfied, not_met]}
              source: {type: string, enum: [static, jev]}
              confidence: {type: [number, 'null'], minimum: 0, maximum: 100}
              impact: {type: string}
...
    Dashboard:
      type: object
      additionalProperties: false
      required: [uniquePassed, byDifficulty, byWeekday, calendar]
      properties:
        uniquePassed: {type: integer, minimum: 0}
        byDifficulty: {type: object, additionalProperties: {type: integer, minimum: 0}}
        byWeekday: {type: object, additionalProperties: {type: integer, minimum: 0}}
        calendar:
          type: array
          items:
            type: object
            required: [publicationDate, passedProblemIds]
            properties:
              publicationDate: {type: string, format: date}
              passedProblemIds: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/Ulid'}}
...
    NoticePage:
      type: object
      required: [items, unreadCount]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, versionId, title, body, publishedAt, read]
            properties:
              id: {$ref: '#/components/schemas/Ulid'}
              versionId: {$ref: '#/components/schemas/Ulid'}
              title: {type: string}
              body: {type: string}
              publishedAt: {type: string, format: date-time}
              read: {type: boolean}
        unreadCount: {type: integer, minimum: 0}
        nextCursor: {type: [string, 'null']}
...
    ProblemError:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}
        detail: {type: string, maxLength: 1000}
        requestId: {type: string}
        retryable: {type: boolean}
        retryAfterSeconds: {type: integer, minimum: 1, maximum: 86400}